+++ app/_components/pdf-sidebar/pdf-sidebar-list-item.tsx
